+++ src/hardware/serialport/libserial.h
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2002-2019  The DOSBox Team
+ *  Copyright (C) 2002-2020  The DOSBox Team
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -16,12 +16,13 @@
  *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
+#include <string.h>
 
 typedef struct _COMPORT *COMPORT;
 
 bool SERIAL_open(const char* portname, COMPORT* port);
 void SERIAL_close(COMPORT port);
-void SERIAL_getErrorString(char* buffer, int length);
+void SERIAL_getErrorString(char* buffer, size_t length);
 
 #define SERIAL_1STOP 1
 #define SERIAL_2STOP 2
